requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# PDF Processing
pdfplumber==0.10.3
PyPDF2==3.0.1

# AI & Vector DB
openai==1.12.0
pinecone-client[grpc]==2.2.4

# HTTP & Utils
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.1
